Move all PCLConfigPlugins sources to `plugins`
This is a follow up to #39649.

In that PR, some shared code in the `plugins` directory of
CondCore/PCLConfigPlugins got factored out and was put into a new header
file that was placed in the `interface` directory.

That resulted in some inconsistencies in the build configuration,
because the `interface` and `src` directories actually didn't correspond
to a normal library, but also to a plugins library because of the
`<flags EDM_PLUGIN="1"/>` in the BuildFile!

Hence, one plugin library included the header files from another plugin
library, which should not be done. In particular, it confused my unused
dependency checker, because you can't declare another plugins library as
a dependency. That forced us to declare the dependencies of the
`CondCore/PCLConfigPlugins` plugins library also in
`CondCore/PCLConfigPlugins`, which caused false positives in the
unused-dependency checks.

This commit fixes the situation by moving the header file into the
`plugins` directory. To avoid confusing a plugin library with a regular
library in the future, the plugin library that was defined in the
top-level of `CondCore/PCLConfigPlugins` got also moved to be another
build target in the `plugins` subdirectory, in a way such that the
compiled library name doesn't change.
